New features
Added a top bar menu, with About, How to read? and Exit buttons.
Created the About window that opens from the menu, which has a description of the Berlin Clock. As well, added a refresh method and button which exits and reopens the Python process for situations that the Python process was hanging (after computer hibernation for example) and the time is not correct.